# backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Contrôle Parental API'
  description: "API REST pour l'application de Contrôle Parental - gestion des familles, enfants, appareils, règles de filtrage, temps d'écran, alertes et rapports."
  version: 1.0.0
servers:
  - url: 'http://localhost:8000'
    description: 'Serveur local'
paths:
...
  /api/families:
    get:
      tags:
        - Families
      summary: 'Liste des familles'
      description: "Retourne les familles de l'utilisateur authentifié (paginées)"
      operationId: familiesIndex
      parameters:
        -
          name: page
          in: query
          description: 'Numéro de page'
          schema:
            type: integer
            default: 1
      responses:
        200:
          description: 'Liste paginée des familles'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Liste des familles'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Family'
                  meta:
                    type: object
                  links:
                    type: object
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          sanctum: []
    post:
      tags:
        - Families
      summary: 'Créer une famille'
      operationId: familiesStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  example: 'Famille Dupont'
                plan:
                  type: string
                  example: free
                  enum:
                    - free
                    - premium
              type: object
      responses:
        201:
          description: 'Famille créée'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Famille créée'
                  data:
                    $ref: '#/components/schemas/Family'
                type: object
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          sanctum: []
  '/api/families/{family}':
    get:
      tags:
        - Families
      summary: "Détails d'une famille"
      operationId: familiesShow
      parameters:
        -
          name: family
          in: path
          description: 'ID de la famille'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Détails de la famille'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Détails de la famille'
                  data:
                    $ref: '#/components/schemas/Family'
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
    put:
      tags:
        - Families
      summary: 'Mettre à jour une famille'
      operationId: familiesUpdate
      parameters:
        -
          name: family
          in: path
          description: 'ID de la famille'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: 'Famille Dupont'
                plan:
                  type: string
                  example: premium
                  enum:
                    - free
                    - premium
              type: object
      responses:
        200:
          description: 'Famille mise à jour'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Famille mise à jour'
                  data:
                    $ref: '#/components/schemas/Family'
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Families
      summary: 'Supprimer une famille'
      description: 'Seul le propriétaire peut supprimer la famille'
      operationId: familiesDestroy
      parameters:
        -
          name: family
          in: path
          description: 'ID de la famille'
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
  /api/locations:
    get:
      tags:
        - Locations
      summary: 'Liste des localisations'
      description: 'Retourne les localisations (paginées), filtrable par child_id'
      operationId: locationsIndex
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
        -
          name: child_id
          in: query
          schema:
            type: integer
      responses:
        200:
          description: 'Liste paginée des localisations'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Liste des localisations'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Location'
                  meta:
                    type: object
                  links:
                    type: object
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          sanctum: []
  '/api/children/{childId}/locations/latest':
    get:
      tags:
        - Locations
      summary: "Dernière localisation d'un enfant"
      operationId: locationsLatest
      parameters:
        -
          name: childId
          in: path
          description: "ID de l'enfant"
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Dernière localisation'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Dernière localisation'
                  data:
                    $ref: '#/components/schemas/Location'
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
...
components:
  schemas:
    User:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Jean Dupont'
        email:
          type: string
          format: email
          example: jean@exemple.com
        phone:
          type: string
          example: '+33612345678'
        avatar:
          type: string
          nullable: true
        locale:
          type: string
          example: fr
        timezone:
          type: string
          nullable: true
        status:
          type: string
          example: active
          enum:
            - active
            - suspended
            - pending
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    Family:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Famille Dupont'
        plan:
          type: string
          example: free
          enum:
            - free
            - premium
        owner:
          $ref: '#/components/schemas/User'
        children_count:
          type: integer
          example: 2
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
    Location:
      properties:
        id:
          type: integer
          example: 1
        child_id:
          type: integer
          example: 1
        device_id:
          type: integer
          example: 1
        latitude:
          type: number
          format: float
          example: 48.8566
        longitude:
          type: number
          format: float
          example: 2.3522
        accuracy_meters:
          type: number
          example: 10.5
          nullable: true
        recorded_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      type: object
...
  responses:
    Success:
      description: 'Opération réussie'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: 'Opération réussie'
              data:
                type: object
                nullable: true
            type: object
    Unauthorized:
      description: 'Non authentifié'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthenticated.
            type: object
    ValidationError:
      description: 'Erreur de validation'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Validation error'
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
            type: object
    NotFound:
      description: 'Ressource introuvable'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Ressource introuvable'
            type: object
    Forbidden:
      description: 'Action non autorisée'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Action non autorisée'
            type: object
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header
...
security:
  -
    sanctum: []